Separated into Query Data and Manage Data
