+++ src/App.jsx
@@ -6,6 +6,7 @@ import Logout from './components/Logout';
 import Layout from './components/Layout/Layout';
 import Dashboard from './pages/Dashboard/Dashboard';
 import Profile from './pages/Profile/Profile';
+import QueryData from './pages/QueryData/QueryData';
 import ManageData from './pages/ManageData/ManageData';
 import AddData from './pages/AddData/AddData';
 import ProtectedRoute from './utils/ProtectedRoute';
@@ -13,6 +14,7 @@ import RegisterPage from './pages/Register/RegisterPage';
 import ManageUsersPage from './pages/ManageUsers/ManageUsersPage';
 import AUTH_ROLES from './utils/auth_config';
 import './module.css';
+// import LoadingScreen from './common/LoadingScreen/LoadingScreen';
 
 const { ADMIN_ROLE, VIEWER_ROLE, EDITOR_ROLE } = AUTH_ROLES.AUTH_ROLES;
 
@@ -59,6 +61,17 @@ function App() {
                 <ProtectedRoute
                   // TODO: Replace ManageData with ManageAttributes
                   Component={ManageData}
+                  redirectPath="/login"
+                  roles={[ADMIN_ROLE, VIEWER_ROLE, EDITOR_ROLE]}
+                />
+              }
+            />
+            <Route
+              exact
+              path="/query-data"
+              element={
+                <ProtectedRoute
+                  Component={QueryData}
                   redirectPath="/login"
                   roles={[ADMIN_ROLE, VIEWER_ROLE, EDITOR_ROLE]}
                 />
