+++ src/common/LoadingScreen/LoadingScreen.module.css
@@ -2,8 +2,9 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 100vw;
-  height: 100vh;
+  width: auto;
+  margin-top: 25%;
+  height: auto;
 }
 
 .content {
