+++ src/pages/ManageData/ManageData.module.css
@@ -1,16 +1,11 @@
 /* Entire container */
-.query-data-container {
+.manage-data-container {
+  /* display: flex;
+  flex-direction: column; */
   padding: 0 1rem 0 1rem;
 }
 
-/* Query Results */
-.query-results-header {
-  margin: 2.5rem 0 2rem 0;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: flex-start;
-}
-.query-results-header h2 {
-  margin: 0;
+.cascades {
+  margin-top: 20px;
+  width: 50%;
 }
