+++ src/pages/QueryData/QueryData.module.css
@@ -1,0 +1,28 @@
+/* Entire container */
+.query-data-container {
+  /* display: flex;
+  flex-direction: column; */
+  padding: 0 1rem 0 1rem;
+}
+
+/* Search type selector */
+.search-type-selector {
+}
+
+/* Generic search bar */
+.generic-search-bar {
+  margin-top: 20px;
+  width: 50%;
+}
+
+/* Query Results */
+.query-results-header {
+  margin: 2.5rem 0 2rem 0;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: flex-start;
+}
+.query-results-header h2 {
+  margin: 0;
+}
