+++ src/pages/QueryData/SelectAttributesModal/AttributeGroup/AttributeGroup.css
