+++ src/pages/QueryData/SelectAttributesModal/AttributeGroup/AttributeGroup.module.css
